binance_bot/requirements.txt
ccxt>=4.0.0
binance-connector>=3.0.0
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
